core/plugins/org.csstudio.dal2.epics/src/org/csstudio/dal2/epics/service/EpicsPvAccessFactory.java
/*
 * Copyright (c) 2012 Stiftung Deutsches Elektronen-Synchrotron,
 * Member of the Helmholtz Association, (DESY), HAMBURG, GERMANY.
 *
 * THIS SOFTWARE IS PROVIDED UNDER THIS LICENSE ON AN "../AS IS" BASIS.
 * WITHOUT WARRANTY OF ANY KIND, EXPRESSED OR IMPLIED, INCLUDING BUT NOT LIMITED
 * TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR PARTICULAR PURPOSE AND
 * NON-INFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
 * FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
 * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR
 * THE USE OR OTHER DEALINGS IN THE SOFTWARE. SHOULD THE SOFTWARE PROVE DEFECTIVE
 * IN ANY RESPECT, THE USER ASSUMES THE COST OF ANY NECESSARY SERVICING, REPAIR OR
 * CORRECTION. THIS DISCLAIMER OF WARRANTY CONSTITUTES AN ESSENTIAL PART OF THIS LICENSE.
 * NO USE OF ANY SOFTWARE IS AUTHORIZED HEREUNDER EXCEPT UNDER THIS DISCLAIMER.
 * DESY HAS NO OBLIGATION TO PROVIDE MAINTENANCE, SUPPORT, UPDATES, ENHANCEMENTS,
 * OR MODIFICATIONS.
 * THE FULL LICENSE SPECIFYING FOR THE SOFTWARE THE REDISTRIBUTION, MODIFICATION,
 * USAGE AND OTHER RIGHTS AND OBLIGATIONS IS INCLUDED WITH THE DISTRIBUTION OF THIS
 * PROJECT IN THE FILE LICENSE.HTML. IF THE LICENSE IS NOT INCLUDED YOU MAY FIND A COPY
 * AT HTTP://WWW.DESY.DE/LEGAL/LICENSE.HTM
 *
 * $Id: DesyKrykCodeTemplates.xml,v 1.7 2010/04/20 11:43:22 bknerr Exp $
 */
package org.csstudio.dal2.epics.service;

import gov.aps.jca.Context;

import org.csstudio.dal2.dv.PvAddress;
import org.csstudio.dal2.dv.Type;
import org.csstudio.dal2.service.DalException;
import org.csstudio.dal2.service.cs.ICsOperationHandle;
import org.csstudio.dal2.service.cs.ICsPvAccess;
import org.csstudio.dal2.service.cs.ICsPvAccessFactory;
import org.csstudio.dal2.service.cs.ICsResponseListener;

/**
 * Factory providing access to the Epics-based pvs
 * 
 * @author jpenning, ascharping
 * @since 07.09.2012
 */
public class EpicsPvAccessFactory implements ICsPvAccessFactory {

	private Context _jcaContext;

	public EpicsPvAccessFactory(final Context jcaContext) {
		assert jcaContext != null : "Precondition: jcaContext != null";
		_jcaContext = jcaContext;
	}

	@Override
	public <T> ICsPvAccess<T> createPVAccess(PvAddress pv, Type<T> type) {
		assert pv != null : "Precondition: pv != null";
		assert type != null : "Precondition: type != null";
		return new EpicsPvAccess<T>(_jcaContext, pv, type);
	}

	@Override
	public ICsOperationHandle requestNativeType(PvAddress pv,
			ICsResponseListener<Type<?>> callback) throws DalException {
		assert pv != null : "Precondition: pv != null";
		assert callback != null : "Precondition: callback != null";
		return new FieldTypeRequester(_jcaContext, pv, callback);
	}

}